move the common llama stack helping function to utils
